[Pal/Linux-SGX,common] Move protected files library to `common`
