+++ common/src/protected_files/lru_cache.c
