+++ common/src/protected_files/meson.build
@@ -1,3 +1,8 @@
+protected_files_inc = [
+    include_directories('.'),
+    common_inc,
+]
+
 protected_files_dep = declare_dependency(
     sources: [
         'lru_cache.c',
@@ -9,10 +14,7 @@ protected_files_dep = declare_dependency(
         'protected_files_internal.h',
     ],
 
-    include_directories: [
-        include_directories('.'),
-        common_inc,
-    ],
+    include_directories: protected_files_inc,
 
     dependencies: [
         uthash_dep,
