+++ common/src/protected_files/protected_files.c
