+++ common/src/protected_files/protected_files_format.h
